requirements.txt
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.29.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-utils==0.7.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
psutil==5.9.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.31.1
